third_party/WebKit/LayoutTests/platform/win/virtual/mojo-loading/http/tests/xmlhttprequest/onloadstart-abort-expected.txt

CONSOLE ERROR: line 51: Uncaught RangeError: Maximum call stack size exceeded.
This is a testharness.js-based test.
PASS Aborting during onloadstart 
PASS Aborting during upload.onloadstart 
PASS Resending during onloadstart 
Harness: the test ran to completion.
